[更新]完善小程序新服务接口
