+++ WeMini/Delivery.php
@@ -17,7 +17,7 @@ namespace WeMini;
 use WeChat\Contracts\BasicWeChat;
 
 /**
- * 小程序退时配送
+ * 小程序即时配送
  * Class Delivery
  * @package WeMini
  */
@@ -174,34 +174,6 @@ class Delivery extends BasicWeChat
     public function reOrder($data)
     {
         $url = 'https://api.weixin.qq.com/cgi-bin/express/local/business/order/readd?access_token=ACCESS_TOKEN';
-        $this->registerApi($url, __FUNCTION__, func_get_args());
-        return $this->callPostApi($url, $data, true);
-    }
-
-    /**
-     * 动态消息，创建被分享动态消息的 activity_id
-     * @param array $data
-     * @return array
-     * @throws \WeChat\Exceptions\InvalidResponseException
-     * @throws \WeChat\Exceptions\LocalCacheException
-     */
-    public function createActivityId($data)
-    {
-        $url = 'https://api.weixin.qq.com/cgi-bin/message/wxopen/activityid/create?access_token=ACCESS_TOKEN';
-        $this->registerApi($url, __FUNCTION__, func_get_args());
-        return $this->callPostApi($url, $data, true);
-    }
-
-    /**
-     * 动态消息，修改被分享的动态消息
-     * @param array $data
-     * @return array
-     * @throws \WeChat\Exceptions\InvalidResponseException
-     * @throws \WeChat\Exceptions\LocalCacheException
-     */
-    public function setUpdatableMsg($data)
-    {
-        $url = 'https://api.weixin.qq.com/cgi-bin/message/wxopen/updatablemsg/send?access_token=ACCESS_TOKEN';
         $this->registerApi($url, __FUNCTION__, func_get_args());
         return $this->callPostApi($url, $data, true);
     }
