+++ WeMini/crypt/wxBizDataCrypt.php
@@ -1,12 +1,10 @@
 <?php
 
 /**
- * 对微信小程序用户加密数据的解密示例代码.
- *
+ * 对微信小程序用户加密数据的解密示例代码
+ * Class WXBizDataCrypt
  * @copyright Copyright (c) 1998-2014 Tencent Inc.
  */
-
-
 class WXBizDataCrypt
 {
     private $appid;
